Add files via upload
A matrix is the rectangular array of numbers. This is a program to input and display a square matrix of size n x n, where the number of rows and columns of the matrix are equal. We can also find the find the sum of main diagonal of the square matrix.